prim_ops.go: 95%

+++ decode.go
@@ -444,7 +444,7 @@ Writes a byte value to an absolute memory location.
 
 NOTE: Do not inline this function as (*sys_wrX) is already inline!
 ****************************************************************************/
-func store_data_byte_abs(segment uint16, offset uint32, val uint8) {
+func store_data_byte_abs(segment uint16, offset uint16, val uint8) {
 	sysw((uint32(segment)<<4 + uint32(offset)), val)
 }
